279 move PropertyNameAttribute to separate namespace
